requirements.txt
asgiref==3.8.1
Django==5.1.6
django-phonenumber-field==8.0.0
phonenumbers==8.13.55
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
